% Auto-generated: do not edit by hand
\name{''DashCookies}

\alias{''DashCookies}

\title{DashCookies component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''DashCookies(id=NULL, data=NULL, max_age=NULL)
}

\arguments{
\item{id}{Character. }

\item{data}{Character | numeric | unnamed list | named list. }

\item{max_age}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

